Add VisAD-based RealTransforms
This is a first cut at two transforms, one based on GriddedDoubleSet,
and the other based on IrregularSet. (VisAD does not have an
IrregularDoubleSet type hierarchy, so we fake the support by doubles
through casts.) Both transforms purport to be invertible because VisAD
supports bidirectional conversion between grid/index space and value
space.